// portfolio-backend/src/app/modules/Projects/project.interface.ts
export interface TProject {

        name: string;
        description: string;
        technologies: string[];
        link: string;
        github: string;
        completionDate: Date;
        image: string[]; 
        status: "Completed" | "In Progress" | "Pending"; 
        features: string[];
        challenges: string[];
        opinions: string[];
      }
      
    